refactor: move api routes and handlers under `./src/api/v1` This refactor is for future proofing.
For adding a new version of API a new folder can be created `./src/api/v2`.
